+++ lib/knossos/solver.rb
@@ -1,3 +1,4 @@
+require_relative 'solver/dijkstra'
 require_relative 'solver/distances'
 
 module Knossos